refactor(cli): split validation helpers from types
