+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-22  
-> **305** tracked files
+> **306** tracked files
 
 ## Summary
 
@@ -14,7 +14,7 @@
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
-| `cli/` | 10 |
+| `cli/` | 11 |
 | `config/` | 9 |
 | `schemas/` | 3 |
 | `blueprints/` | 2 |
@@ -30,7 +30,7 @@
 | `.md` | 90 |
 | `.go` | 53 |
 | `.yaml` | 23 |
-| `.rs` | 8 |
+| `.rs` | 9 |
 | `.yml` | 6 |
 | `(no extension)` | 5 |
 | `.toml` | 4 |
@@ -159,7 +159,8 @@
 │       │   ├── validate.rs
 │       │   └── workflow.rs
 │       ├── main.rs
-│       └── types.rs
+│       ├── types.rs
+│       └── validation.rs
 ├── cliff.toml
 ├── cmd/
 │   └── orchestrator/
